Fix handling of changes to single-line files
For some reason the syntax for a change to a single-line file was
conflated with the syntax for adding or deleting a file. This caused a
diff like
```diff
diff --git a/.nvmrc b/.nvmrc
--- a/.nvmrc
+++ b/.nvmrc
@@ -1 +1 @@
-14.16.1
+16.7.0
```
to be interpreted as deleting the file rather than changing it.

Fixing that involves deleting the erroneous constants
FILE_DELETION_LINE_COUNT and FILE_CREATION_LINE_COUNT, and changing all
the logic that used them to correctly interpret their replacements.

I note the logic here still isn't quite correct, as it does not
differentiate between a file that is deleted and one that is truncated
to 0 bytes, or one that is created and one that was formerly 0 bytes.
I'll leave that for someone else to figure out; it seems the general
answer is to look at the timestamp, but svn doesn't output any
timestamps.

I also noted that the behavior with respect to filenames reported for
added and deleted files was inconsistent: if the file had multiple
lines the original and new name in the diff header was reported, while
if it had a single line the unnecessary name was omitted. This PR makes
it consistently use the former behavior, as that seems like the most
common case that users of the library would have encountered.